combine memory tests per GC. Fix bug in GC duration tests to account for simulteneous young and old gc.
