--- jfr-streaming/src/serialGcTest/java/io/opentelemetry/contrib/jfr/metrics/SerialMemoryLimitMetricTest.java
@@ -1,15 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.contrib.jfr.metrics;
-
-import org.junit.jupiter.api.Test;
-
-class SerialMemoryLimitMetricTest extends AbstractMetricsTest {
-  @Test
-  void shouldHaveMemoryLimitMetrics() {
-    // TODO: needs JFR support. Placeholder.
-  }
-}
